+++ wezterm/src/main.rs
@@ -6,6 +6,13 @@ use config::{wezterm_version, ConfigHandle};
 use mux::Mux;
 use std::ffi::OsString;
 use std::io::Read;
+use termwiz::caps::Capabilities;
+use termwiz::escape::esc::{Esc, EscCode};
+use termwiz::escape::OneBased;
+use termwiz::input::{InputEvent, KeyCode, KeyEvent};
+use termwiz::surface::change::Change;
+use termwiz::surface::Position;
+use termwiz::terminal::{ScreenSize, Terminal};
 use umask::UmaskSaver;
 use wezterm_gui_subcommands::*;
 
@@ -153,13 +160,147 @@ struct ImgCatCommand {
     /// ratio
     #[arg(long = "no-preserve-aspect-ratio")]
     no_preserve_aspect_ratio: bool,
+
+    /// Set the cursor position prior to displaying the image.
+    /// The default is to use the current cursor position.
+    /// Coordinates are expressed in cells with 0,0 being the top left
+    /// cell position.
+    #[arg(long, value_parser=ValueParser::new(x_comma_y))]
+    position: Option<ImagePosition>,
+
+    /// Do not move the cursor after displaying the image.
+    /// Note that when used like this from the shell, there is a very
+    /// high chance that shell prompt will overwrite the image;
+    /// you may wish to also use `--hold` in that case.
+    #[arg(long)]
+    no_move_cursor: bool,
+
+    /// Wait for enter to be pressed after displaying the image
+    #[arg(long)]
+    hold: bool,
+
+    /// How to manage passing the escape through to tmux
+    #[arg(long, value_parser)]
+    tmux_passthru: Option<TmuxPassthru>,
+
     /// The name of the image file to be displayed.
     /// If omitted, will attempt to read it from stdin.
     #[arg(value_parser, value_hint=ValueHint::FilePath)]
     file_name: Option<OsString>,
 }
 
+#[derive(Clone, Copy, Debug)]
+struct ImagePosition {
+    x: u32,
+    y: u32,
+}
+
+fn x_comma_y(arg: &str) -> Result<ImagePosition, String> {
+    if let Some(eq) = arg.find(',') {
+        let (left, right) = arg.split_at(eq);
+        let x = left.parse().map_err(|err| {
+            format!("Expected x,y to be integer values, got {arg}. '{left}': {err:#}")
+        })?;
+        let y = right[1..].parse().map_err(|err| {
+            format!("Expected x,y to be integer values, got {arg}. '{right}': {err:#}")
+        })?;
+        Ok(ImagePosition { x, y })
+    } else {
+        Err(format!("Expected name=value, but got {}", arg))
+    }
+}
+
+#[derive(Debug)]
+pub(crate) struct ImageInfo {
+    pub width: u32,
+    pub height: u32,
+    pub _format: image::ImageFormat,
+}
+
 impl ImgCatCommand {
+    fn compute_image_cell_dimensions(
+        &self,
+        info: ImageInfo,
+        term_size: ScreenSize,
+    ) -> (usize, usize) {
+        let physical_cols = term_size.cols;
+        let physical_rows = term_size.rows;
+        let cell_pixel_width = term_size.xpixel;
+        let cell_pixel_height = term_size.ypixel;
+        let pixel_width = cell_pixel_width * physical_cols;
+        let pixel_height = cell_pixel_height * physical_rows;
+
+        let width = self
+            .width
+            .unwrap_or_default()
+            .to_pixels(cell_pixel_width, physical_cols);
+        let height = self
+            .height
+            .unwrap_or_default()
+            .to_pixels(cell_pixel_height, physical_rows);
+
+        // Compute any Automatic dimensions
+        let aspect = info.width as f32 / info.height as f32;
+
+        let (width, height) = match (width, height) {
+            (None, None) => {
+                // Take the image's native size
+                let width = info.width as usize;
+                let height = info.height as usize;
+                // but ensure that it fits
+                if width as usize > pixel_width || height as usize > pixel_height {
+                    let width = width as f32;
+                    let height = height as f32;
+                    let mut candidates = vec![];
+
+                    let x_scale = pixel_width as f32 / width;
+                    if height * x_scale <= pixel_height as f32 {
+                        candidates.push((pixel_width, (height * x_scale) as usize));
+                    }
+                    let y_scale = pixel_height as f32 / height;
+                    if width * y_scale <= pixel_width as f32 {
+                        candidates.push(((width * y_scale) as usize, pixel_height));
+                    }
+
+                    candidates.sort_by(|a, b| (a.0 * a.1).cmp(&(b.0 * b.1)));
+
+                    candidates.pop().unwrap()
+                } else {
+                    (width, height)
+                }
+            }
+            (Some(w), None) => {
+                let h = w as f32 / aspect;
+                (w, h as usize)
+            }
+            (None, Some(h)) => {
+                let w = h as f32 * aspect;
+                (w as usize, h)
+            }
+            (Some(w), Some(_)) if !self.no_preserve_aspect_ratio => {
+                let h = w as f32 / aspect;
+                (w, h as usize)
+            }
+            (Some(w), Some(h)) => (w, h),
+        };
+
+        // And convert to cells
+        (width / cell_pixel_width, height / cell_pixel_height)
+    }
+
+    fn image_dimensions(data: &[u8]) -> anyhow::Result<ImageInfo> {
+        let reader = image::io::Reader::new(std::io::Cursor::new(data)).with_guessed_format()?;
+        let format = reader
+            .format()
+            .ok_or_else(|| anyhow::anyhow!("unknown format!?"))?;
+        let (width, height) = reader.into_dimensions()?;
+        Ok(ImageInfo {
+            width,
+            height,
+            _format: format,
+        })
+    }
+
     fn run(&self) -> anyhow::Result<()> {
         let mut data = Vec::new();
         if let Some(file_name) = self.file_name.as_ref() {
@@ -171,6 +312,69 @@ impl ImgCatCommand {
             stdin.read_to_end(&mut data)?;
         }
 
+        let caps = Capabilities::new_from_env()?;
+        let mut term = termwiz::terminal::new_terminal(caps)?;
+        term.set_raw_mode()?;
+
+        let mut probe = term
+            .probe_capabilities()
+            .ok_or_else(|| anyhow!("Terminal has no prober?"))?;
+
+        let xt_version = probe.xt_version()?;
+
+        let term_size = probe.screen_size()?;
+
+        let is_tmux = xt_version.is_tmux();
+
+        // TODO: ideally we'd do some kind of probing to see if conpty
+        // is in the mix. For now we just assume that if we are on windows
+        // then it must be in there somewhere.
+        let is_conpty = cfg!(windows);
+
+        // Not all systems understand that the cursor should move as
+        // part of processing the image escapes, so we need to move it
+        // explicitly after we've drawn things.
+        // We can only do this reasonably sanely if we aren't setting
+        // the absolute position.
+        let needs_force_cursor_move = !self.no_move_cursor && !self.position.is_some() && (is_tmux || is_conpty)
+            // We can only use forced movement if we know the pixel geometry
+            && (term_size.xpixel != 0 && term_size.ypixel != 0);
+
+        term.set_cooked_mode()?;
+
+        let save_cursor = Esc::Code(EscCode::DecSaveCursorPosition);
+        let restore_cursor = Esc::Code(EscCode::DecRestoreCursorPosition);
+
+        if let Some(position) = &self.position {
+            let csi = termwiz::escape::CSI::Cursor(
+                termwiz::escape::csi::Cursor::CharacterAndLinePosition {
+                    col: OneBased::from_zero_based(position.x),
+                    line: OneBased::from_zero_based(position.y),
+                },
+            );
+            print!("{save_cursor}{csi}");
+        }
+
+        let (begin, end) = self.tmux_passthru.unwrap_or_default().get();
+
+        let image_dims = Self::image_dimensions(&data)
+            .map(|info| self.compute_image_cell_dimensions(info, term_size));
+
+        if let (Ok((_cursor_x, cursor_y)), true) = (&image_dims, needs_force_cursor_move) {
+            // Before we emit the image, we need to emit some new lines so that
+            // if the image would scroll the display, things end up in the right place
+            let new_lines = "\n".repeat(*cursor_y);
+            print!("{new_lines}");
+
+            // and move back up again.
+            // Note that we lose the x cursor position and land back in the first
+            // column as a result of doing this.
+            term.render(&[Change::CursorPosition {
+                x: Position::Absolute(0),
+                y: Position::Relative(-1 * (*cursor_y as isize)),
+            }])?;
+        }
+
         let osc = OperatingSystemCommand::ITermProprietary(ITermProprietary::File(Box::new(
             ITermFileData {
                 name: None,
@@ -179,11 +383,36 @@ impl ImgCatCommand {
                 height: self.height.unwrap_or_default(),
                 preserve_aspect_ratio: !self.no_preserve_aspect_ratio,
                 inline: true,
-                do_not_move_cursor: false,
+                do_not_move_cursor: self.no_move_cursor,
                 data,
             },
         )));
-        println!("{}", osc);
+        println!("{begin}{osc}{end}");
+
+        if let (Ok((_cursor_x, cursor_y)), true) = (&image_dims, needs_force_cursor_move) {
+            // tell the terminal that doesn't fully understand the image sequence
+            // to move the cursor to where it should end up
+            term.render(&[Change::CursorPosition {
+                x: Position::Absolute(0),
+                y: Position::Relative(*cursor_y as isize),
+            }])?;
+        } else if self.position.is_some() {
+            print!("{restore_cursor}");
+        }
+
+        if self.hold {
+            while let Ok(Some(event)) = term.poll_input(None) {
+                match event {
+                    InputEvent::Key(KeyEvent {
+                        key: KeyCode::Enter,
+                        modifiers: _,
+                    }) => {
+                        break;
+                    }
+                    _ => {}
+                }
+            }
+        }
 
         Ok(())
     }
@@ -195,6 +424,10 @@ struct SetCwdCommand {
     /// If omitted, will use the current directory of the process itself.
     #[arg(value_parser, value_hint=ValueHint::DirPath)]
     cwd: Option<OsString>,
+
+    /// How to manage passing the escape through to tmux
+    #[arg(long, value_parser)]
+    tmux_passthru: Option<TmuxPassthru>,
 
     /// The hostname to use in the constructed file:// URL.
     /// If omitted, the system hostname will be used.
@@ -218,9 +451,44 @@ impl SetCwdCommand {
         let host = host.to_str().unwrap_or("localhost");
         url.set_host(Some(host))?;
 
+        let (begin, end) = self.tmux_passthru.unwrap_or_default().get();
+
         let osc = OperatingSystemCommand::CurrentWorkingDirectory(url.into());
-        print!("{}", osc);
+        print!("{begin}{osc}{end}");
+        if !begin.is_empty() {
+            // Tmux understands OSC 7 but won't automatically pass it through.
+            // <https://github.com/tmux/tmux/issues/3127#issuecomment-1076300455>
+            // Let's do it again explicitly now.
+            print!("{osc}");
+        }
         Ok(())
+    }
+}
+
+#[derive(Copy, Clone, Debug, ValueEnum, Default)]
+enum TmuxPassthru {
+    Disable,
+    Enable,
+    #[default]
+    Detect,
+}
+
+impl TmuxPassthru {
+    fn is_tmux() -> bool {
+        std::env::var_os("TMUX").is_some()
+    }
+
+    fn get(&self) -> (&'static str, &'static str) {
+        let enabled = match self {
+            Self::Enable => true,
+            Self::Detect => Self::is_tmux(),
+            Self::Disable => false,
+        };
+        if enabled {
+            ("\u{1b}Ptmux;\u{1b}", "\u{1b}\\")
+        } else {
+            ("", "")
+        }
     }
 }